wybór specjalności, wybór ubezpieczenia, edycja danych

+++ src/main/java/edu/uph/ii/platformy/repositories/UbezpieczenieRepository.java
@@ -5,7 +5,7 @@ import edu.uph.ii.platformy.models.Kierunki;
 import edu.uph.ii.platformy.models.Ubezpieczenie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UbezpieczenieRepository extends JpaRepository<Ubezpieczenie, Integer> {
+public interface UbezpieczenieRepository extends JpaRepository<Ubezpieczenie, Long> {
 
     Ubezpieczenie findByName(String name);
 }
